requirements.txt
opencv-python~=4.8.0.76
opencv-contrib-python==4.8.0.76
Pillow==10.0.1
ultralytics~=8.0.180
numpy~=1.24.3
paho-mqtt~=1.6.1
